Commit 23.06
Added: Hercules Gunship special enemy

Improved: Player controlability of the aircraft, behavior of the aircraft
Improved: Settings for Aim Sensitivity and Y axis reversal.

Implemented: Autofire mechanic

+++ Assets/Scripts/Weapons/Gun.cs
@@ -22,7 +22,7 @@ public class Gun : MonoBehaviour
     [SerializeField] float overheatResetTimer;
     [SerializeField] float overheatedRoF;
 	
-	public float shotsFired;
+	public int shotsFired;
 
     [SerializeField] KillCounter killCounter;
